+++ src/@configs/navigation.js
@@ -32,6 +32,7 @@ import { BsFillMarkdownFill     as MarkdownIcon       } from 'react-icons/bs' ;
 import { GiDominoMask           as MaskIcon           } from "react-icons/gi";
 import { RiLayoutMasonryFill    as MasonryIcon        } from "react-icons/ri";
 import { TfiLayoutMenuSeparated as MenuIcon           } from "react-icons/tfi" ;
+import { LuLayoutPanelTop       as MegamenuIcon       } from "react-icons/lu";
 import { SiDialogflow           as ModalIcon          } from "react-icons/si";
 import { MdAnimation            as MotionIcon         } from "react-icons/md";
 import { BiFoodMenu             as NavigationIcon     } from "react-icons/bi";
@@ -112,6 +113,7 @@ const navigation =
                 [
                     { id : 'pagination' , type : LINK  , Icon : PaginationIcon , path  : '/lab/pagination' } ,
                     { id : 'menus'      , type : LINK  , Icon : MenuIcon       , path  : '/lab/menus'      } ,
+                    { id : 'megamenu'   , type : LINK  , Icon : MegamenuIcon   , path  : '/lab/megamenu'   } ,
                 ]
             } ,
             {